+++ tests/lib/Integration/Eloquent/HasManyThroughTest.php
@@ -17,6 +17,7 @@
 
 namespace CloudCreativity\LaravelJsonApi\Tests\Integration\Eloquent;
 
+use CloudCreativity\LaravelJsonApi\Tests\Integration\TestCase;
 use DummyApp\Country;
 use DummyApp\Post;
 use DummyApp\User;